fix: restore ROCm BF16 inference correctness

+++ src/backend.rs
@@ -4,7 +4,7 @@ use burn::tensor::backend::Backend;
 pub type CpuBackend = burn_ndarray::NdArray<f32, i32, i8>;
 
 #[cfg(feature = "rocm")]
-pub type RocmBackend = burn_rocm::Rocm<f32, i32, u8>;
+pub type RocmBackend = burn_rocm::Rocm<half::bf16, i32, u8>;
 
 pub trait NanoBackend: Backend<IntElem = i32> {}
